Implement larger pool of shape and color options
Anything that produces an svg should more or less be fine to clip to the
bounds of a shape. This lets us define color as a generic blob of svg
objects, and clip it using shape, which is also a generic blob of svg
objects. This gets us into a convenient state of being able to combine
voronoi + triangle, for example.

Because of the new options, also add a new, second custom page as a
sandbox.